Move task_registry to core
